Add footer to messages sent
Add a message header indicating the version of the program used.

Add footer to message showing it was sent using the Python Bulk Email
script.  This can be disabled through a setting in the configuration
file or by including the --no-footer command-line option.

Fix typo on --addr-file command line argument.